pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = [
    {name = "jmhwang-dev",email = "jungmin.hwang.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "duckdb (>=1.3.1,<2.0.0)",
    "kaggle (>=1.7.4.5,<2.0.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "transformers (==4.51.3)",
    "torch (==2.6.0)",
    "accelerate (==1.6.0)",
    "emoji (>=2.14.1,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "wordcloud (>=1.9.4,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "notebook (>=7.4.4,<8.0.0)",
    "pydantic (==2.11.7)",
    "kafka-python (==2.2.15)",
    "fastapi (==0.116.1)",
    "uvicorn (==0.35.0)",
    "fastavro (>=1.11.1,<2.0.0)",
    "confluent-kafka[avro] (>=2.11.0,<3.0.0)",
    "optimum[onnxruntime] (>=1.27.0,<2.0.0)",
    "onnxruntime-gpu (>=1.22.0,<2.0.0)",
    "boto3 (>=1.40.3,<2.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyyml (>=0.0.2,<0.0.3)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  { include = "ecommerce", from = "src" },
  { include = "service", from = "src" }
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.4.1"
pyspark = "^3.5"
notebook = "^7.4.4"

